+++ python/paddle/fluid/layers/nn.py
@@ -9829,7 +9829,7 @@ def size(input):
     """
     **Size Layer**
 
-    Returns the number of elements for a tensor, which is a int32 Tensor with shape [1].
+    Returns the number of elements for a tensor, which is a int64 Tensor with shape [1].
 
     Args:
         input (Variable): The input variable.
@@ -9846,7 +9846,7 @@ def size(input):
     """
 
     helper = LayerHelper('size', **locals())
-    out = helper.create_variable_for_type_inference(dtype='int32')
+    out = helper.create_variable_for_type_inference(dtype='int64')
     helper.append_op(type='size', inputs={'Input': input}, outputs={'Out': out})
 
     return out
